requirements.txt
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
crewai>=0.100.0
langchain-ollama>=0.2.0
crewai-tools>=0.12.0
python-dotenv==1.0.1
pysqlite3-binary
pdfplumber
pypdf
python-docx
openpyxl

# Servidor web
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
markdown>=3.6